adk_agent/pyproject.toml
[project]
name = "adk-agent"
version = "0.1.0"
description = "A2A ADK Agent Implementation"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Dependencies specific to ADK agent itself
    "pyyaml",
    "uvicorn",
    "starlette",
    "sse-starlette",
    "pydantic",
    "google-adk",
    # Add dependencies required by the mounted 'common' code
    "httpx>=0.23.0", # Version adjusted for compatibility previously
    "httpx-sse>=0.4.0",
    "jwcrypto>=1.5.6",
    "pyjwt>=2.10.1",
    "typing-extensions>=4.12.2",
]

# [tool.uv.sources] section removed
